+++ src/component/Page.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Navbar from './Navbar';
+import Navbar from '../components/Navbar';
 import Home from '../page/Home';
 import About from '../page/About';
 import Content from '../page/Content';
